services/socialfy/AgenticOSKevsAcademy/requirements.txt
# Core Framework
aiohttp>=3.9.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
redis>=5.0.0
celery>=5.3.0
websockets>=12.0
httpx>=0.25.0
typing-extensions>=4.8.0
dataclasses-json>=0.6.0

# Browser Automation (Instagram DM Agent)
playwright>=1.40.0
playwright-stealth>=1.0.6  # Anti-detection para automação

# Database
psycopg2-binary>=2.9.9
alembic>=1.13.0
supabase>=2.0.0

# Environment & Config
python-dotenv>=1.0.0

# Security & Encryption
cryptography>=41.0.0
PyJWT>=2.8.0  # JWT validation for auth middleware

# HTTP Clients
requests>=2.31.0

# Google Services
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.111.0
gspread>=5.12.0

# Apify
apify-client>=1.6.0

# AI/ML
google-generativeai>=0.3.0
anthropic>=0.8.0
openai>=1.0.0
Pillow>=10.0.0

# Video Processing
yt-dlp>=2023.12.30

# Data Processing
pandas>=2.1.0
python-dateutil>=2.8.0

# Logging
colorlog>=6.8.0

# Monitoring
psutil>=5.9.0

# Timezone
pytz>=2024.1
